Parametrize EArray/EBinary on length encoding
Keep EArray2/EBinary2 apply constructors
